+++ simpleble/src/backends/common/AdapterBase.h
@@ -5,10 +5,10 @@
 #include <string>
 #include <vector>
 
-#include <simpleble/export.h>
-
 #include <simpleble/Exceptions.h>
 #include <simpleble/Types.h>
+
+#include <kvn_safe_callback.hpp>
 
 namespace SimpleBLE {
 
@@ -22,7 +22,6 @@ class PeripheralBase;
  *
  * Notes for implementers:
  *
- * - The `initialize()` method is not present because internally we enforce RAII.
  * - The methods here return shared pointers to AdapterBase, PeripheralBase,
  *   etc. The code in Adapter.cpp will automatically wrap these in Adapter,
  *   Peripheral, etc. objects.
@@ -39,10 +38,11 @@ class AdapterBase {
     virtual std::string identifier() = 0;
     virtual BluetoothAddress address() = 0;
 
-    // NOTE: A default implementation is provided for these methods until all backends have implemented them.
     virtual void power_on() = 0;
     virtual void power_off() = 0;
     virtual bool is_powered() = 0;
+    virtual void set_callback_on_power_on(std::function<void()> on_power_on);
+    virtual void set_callback_on_power_off(std::function<void()> on_power_off);
 
     virtual void scan_start() = 0;
     virtual void scan_stop() = 0;
@@ -67,6 +67,9 @@ class AdapterBase {
 
   protected:
     AdapterBase() = default;
+
+    kvn::safe_callback<void()> _callback_on_power_on;
+    kvn::safe_callback<void()> _callback_on_power_off;
 };
 
 }  // namespace SimpleBLE